pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sys-cli"
version = "0.1.0"
description = "Personal system automation CLI (WSL-first)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1.7",
  "PyYAML>=6.0.1",
]

[project.scripts]
sys = "sys_cli.cli:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["sys_cli*"]
